dev-util/rust-script: check for Cargo config.toml before creation
https://public-inbox.gentoo.org/gentoo-dev/20240725180612.19453-2-chewi@gentoo.org/T/#u
will replace `config` with `config.toml`, thus cause the creation of the
latter to fail.

+++ dev-util/rust-script/rust-script-0.34.0.ebuild
@@ -122,5 +122,7 @@ src_prepare() {
 	use debug || sed -i "s|/debug/|/release/|" tests/util/mod.rs || die
 
 	# cargo.eclass uses the old config and deprecation warning breaks the test.
-	ln -s "${ECARGO_HOME}/config" "${ECARGO_HOME}/config.toml" || die
+	if [[ ! -e "${ECARGO_HOME}/config.toml" ]]; then
+		ln -s "${ECARGO_HOME}/config" "${ECARGO_HOME}/config.toml" || die
+	fi
 }
